New pg_regress argument to ignore plan diffs.
This is a preparation of an upcoming random ICG pipeline job.  In that
pipeline job we will have tables created on different segments, so the
plans might be different expectation.  In atmsort.pm there is already an
argument -gpd_ignore_plans to ignore the plan diffs, but there was no
equivalence in pg_regress.

Added a pg_regress argument --ignore-plans to ignore plan diffs.  It can
be enabled via make command like this:

    make installcheck EXTRA_REGRESS_OPTS=--ignore-plans