Dispatch Alter System
ALTER SYSTEM only modified the GUC on coordinator. Changing it to
dispatch the command to segments as well. Would be ideal in long run
for ALTER SYSTEM similar to gpconfig to have coordinator_only
option. But for now it seems more natural to set GUC on all nodes
instead of only coordinator. Given this command directly works
directly on filesystem and not transactional, if fails on a segment,
the modification will still remain on other segments. I believe this
is no different from gpconfig failing half way.

Reviewed-by: Asim R P <pasim@vmware.com>