Allow reshuffling tables with update triggers
Previously, when updating a table with update triggers on its distribution column,
GPDB report an error like "ERROR: UPDATE on distributed key column not allowed on
relation with update triggers" because current GPDB executor don't support
statement-level update triggers and will also skip row-level update triggers for
a split-update is actually consist of delete and insert, so if the result relation
has update triggers, GPDB reject and error out because it's not functional.

There is an exception for 'ALTER TABLE SET WITH (RESHUFFLE)', RESHUFFLE also use
split-update node internal to rebalance/expand table, however, from the view of
users, ALTER TABLE should not hit any kind of triggers, so we don't need to
error out as same as UPDATE command.