Make gpnetbench Makefile use standard constructs
gpnetbench was building more like a separate project, than as a part
of a project with etablished patterns for how to build.  This ports
the gpnetbench Makefile over to using the standard way to build objs
and binaries in the tree.  Also removes an unused variable from the
Makefile in the process.