Fix Extra Text processing in EXPLAIN
Tidy up and fix various issues with the Extra Text printing in EXPLAIN
processing.

* Avoid ending output with a newline if not needed
* Close the groups in the right order to ensure valid JSON
* Use StringInfoData per how most allocations are done in the code
  and free the correct allocation to actually free the buffer rather
  then leaking it.
* Move the leading whitespace in allocation to where it's actually
  needed rather than always ending with lots of blanks.
* Minor style cleanups

The text is still printed in Text for formats other than text output,
but this patch doesn't make that any worse. Fixing a common way for
Extra Text to make use of the structured formatting is for another
patch.

Reviewed-by: Joao Pereira <jdealmeidapereira@pivotal.io>
Reviewed-by: Ning Yu <nyu@pivotal.io>
Reviewed-by: Venkatesh Raghavan <vraghavan@pivotal.io>