Remove NULL guards from pg_dump to align with upstream
To answer the question raised in the removed FIXME comment: owner is
not allowed to be NULL, it may be omitted if --no-owner it passed but
each ArchiveEntry should have an owner set; dropStmt is allowed to be
empty but not NULL, the pattern for not setting a dropStmt is to pass
"" in the ArchiveEntry call.

While these guards didn't really hurt, every non-essential divergence
from upstream can cause conflicts when merging so we might as well
keep aligned. If we want changes along these lines we should submit
them upstream.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>