Add EXPLAIN format output to MemAccounting
This adds support for the structured formats in EXPLAIN (format ..) to
the MemAccounting code, which previously only supported string output.
The walker code is updated to take the ExplainState and make decisions
on output format rather then just the take the buffer and assume text
output format.

As the PrettyPrinting debug function was using the ToString walker, it
was removed in this commit since there are no callsites actually using
it. The unittest for PrettyPrint was also removed, and the unittest for
ToString updated to match the new code.

Also tidies up the header includes to be in alphabetical order as per
how the code is usually structured.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>