Use Postgres hash functions for MCV merging.
This code is concerned with partitioning, not with distribution across
segments, so there's no need to rely on the "cdb" hashing. The cdb hashing
only works for a limited set of built-in datatypes, so this allows merging
leaf stats for any datatype that has a hash opclass defined.

One small downside of this is that there are actually two built-in datatypes
that have cdbhash support: money and varbit. So we leaf statistics merging
no longer works for them. That's an acceptable loss, those datatypes should
hardly be used in production, anyway.

Another little thing that caused expected output changes was that the new
code makes a different choice on which MCVs to keep, when there is a tie
between multiple values. The choice is arbitrary, and the new one is as
good as the old one.