Get rid of cdbhash usage in predtest.
Use built-in hash_any instead. The "match" function uses equal() to
compare the Consts, which compares the raw bytes instead of doing a
datatype-aware comparison, so we might as well hash the raw bytes in the
hash function, too.

While we're at it, refactor the predicate testing code a bit. Move the
functions related to the "possible values" computation to a separate file,
and clean up and refactor the code for readability.