pg_resetxlog: add option to update system identifier
For upgrade, we need to reset the database system identifier in the
pg_control file for the segments. The --system-identifier option lets us
do that.

Per Daniel's suggestion, move all GPDB-specific additions to use long
options; the -y option is now --binary-upgrade. We've based the new
long-options machinery on e22b27f0c from upstream, which adds long
options to pg_resetwal. This should make it easier to merge later.

Co-authored-by: Jim Doty <jdoty@pivotal.io>
Co-authored-by: Shoaib Lari <slari@pivotal.io>