pg_upgrade: reset segment system identifiers using pg_resetxlog
The copied system identifiers from the master need to be made unique so
that mirrors cannot accidently connect to the wrong primaries.  We do
this with the --system-identifier option to pg_resetxlog introduced in
the last commit.

We have also added a test to verify unique system identifiers are
assigned to the segments.

Co-authored-by: Shoaib Lari <slari@pivotal.io>
Co-authored-by: Jim Doty <jdoty@pivotal.io>