Fix reshuffle issue on inheritance table
When reshuffling inheritance table whose children have different gppolicy,
we find that the policy of children tables are changed to the same policy
with the parent table on all segments, eg, parent table mix_base_tbl is
distributed by (c1), its child mix_child_a is distributed by (c2), after
a reshuffling, the policy of mix_child_a is changed to be distributed by
(c1) on all segments. This is not by designed, reshuffle only expect the
numsegments property changes.