Ensure to return the Oid of the new extension
On the QE's, make sure to keep the API property of CreateExtension()
by returning the Oid of the newly created extension object.  There
are no callsites that use the return value currently, so this has
little to no effect, but in case we start using it we may as well
get it right from the start.

Reviewed-by: BaiShaoqi <sbai@pivotal.io>
Reviewed-by: Melanie Plageman <mplageman@pivotal.io>
Reviewed-by: Jimmy Yih <jyih@pivotal.io>