Allow registering for compaction drop for segment files in compactedSegmentFileList.
When there are many concurrent operations on AO tables, it is possible
the aoentry fetched during RegisterSegnoForCompactionDrop is a brand
new entry that does not contain information about the current vacuum
of the relation. In this case, compactedSegmentFileList contains the
accurate list of segment files that have been compacted.

Remove the elogif that assumes the aoentry is accurate.

Note: The aoentry will not be evicted again after RegisterSegnoForCompactionDrop
because the entry is marked as 'in use'.

Co-authored-by: Asim R P <apraveen@pivotal.io>