Remove unnecessary GPDB code to transform PathKeys of index paths.
As far as I can tell, GPDB works the same as PostgreSQL with regards to
path keys used for append rels, so I don't see why we'd need to do any
transformation here. Regression tests are passing without it.

This code has been moved around as part of the 9.2 merge, and some other
cleanup, but goes all the way back to 2007 in the old pre-open sourcing
repository. The commit that introduced was a massive commit with message
"Merge of Release-3_1_0_0-alpha1-branch branch down to HEAD", so I lost the
trace of its origin there. I guess it was needed back then, but seems
unnecessary now.