ao_upgrade: load GPDB4 numerics via binary format
This test has already been rewritten twice due to upstream format
changes. It's better to just hardcode the numerics we want to test in
the original GPDB4 format (thanks to Heikki for this idea!). As a
result, we no longer need the convert_to_v4() and
numeric_force_long_format() helper functions.

We also remove the GPDB_93_MERGE_FIXME that was added recently. The
reason this test changed output format is that the mangled GPDB4
numerics had extra digits past the display scale. This is not supposed
to happen during normal operation, and in upstream commit 5cb0e335 --
which was merged as part of the 9.3 iteration -- get_str_from_var() was
changed to no longer perform rounding, which leads to truncation for
these "corrupted" numerics.

Co-authored-by: Alexandra Wang <lewang@pivotal.io>