Fix support of recovery_target_timeline recovery.conf setting
When recovery_target_timeline setting is set in recovery.conf, we expect the
recoveryTargetTLI variable to change (e.g. when set to 'latest', it will read
the primary's history file to get latest TLI). This logic seems to have been
removed from the bottom of readRecoveryCommandFile() when WAL replication was
backported for master/standby replication.

Co-authored-by: Paul Guo <pguo@pivotal.io>