Remove early exit in RequestXLogStreaming
When WalRcv's walRcvState is not WALRCV_STOPPED, we do early exit. This early
exit is not in upstream Postgres and is legacy Greenplum code. This prevented
timeline switch scenario from happening correctly because WAL receiver's state
is set to WALRCV_WAITING expecting to wake itself up with SetLatch or
PMSIGNAL_START_WALRECEIVER signal call at the end of RequestXLogStreaming()
call.

Co-authored-by: Paul Guo <pguo@pivotal.io>