Add simple pg_rewind test to test no rewind required scenario
All of the pg_rewind tests currently only test scenarios where pg_rewind finds
divergence and rewinds the target. This test being added will test the scenario
where pg_rewind decides that no rewind is required on the target and that the
target will be able to catch up automatically to the source server's new
timeline.

Co-authored-by: Paul Guo <pguo@pivotal.io>