Make Greenplum tablespaces work with gprecoverseg -F
Greenplum tablespaces are very different from upstream Postgres. It creates a
subdirectory under the tablespace location path. When running pg_basebackup with
existing user-defined tablespaces, it will error out because the tablespace root
path already exists. For Greenplum, it should check the correct subdirectory in
the tablespace path instead.

To do this, we introduce --target-gp-dbid flag to pg_basebackup and introduce
logic for user-defined tablespace. Then we change utilities to use the new option.

Co-authored-by: Jinbao Chen <jinchen@pivotal.io>