Simplify code to choose a segment in random, for DISTRIBUTED RANDOMLY.
The definition of cdb_randint() was pretty hard to understand. Where did
the 0.999999 constant come from, for example? And the signature of
cdb_randint() was also surprising, with the upper bound as the first
argument, and lower bound second. The call in makeRandomSegMap() got that
backwards, although it still mostly worked. Replace cdb_randint() with a
more straightforward cdbhashrandomseg(int numsegments) function.

Also simplify the 'rrindex' mechanism used in cdbhash.c, to choose a
segment in random. The comment in makeCdbHash() that claimed that the
calling cdbhashnokey() repeatedly would behave in a round-robin fashion
was wrong: the 'rrindex' counter that was incremented on every call was
fed to a hash function, which meant that we were using the hash function
as a random number generator. Since we were using it as a random number
generator anyway, remove the 'rrindex' field and call random() directly
in cdbhashnokey().

Fixes https://github.com/greenplum-db/gpdb/issues/5899.

Reviewed-by: Zhenghua Lyu <zlv@pivotal.io>