Rewrite WAL-logging of bitmap AM insertion.
There were a couple of issues with the old code:

1. It was extending the relation, and doing palloc's, while in critical
   section. Those can fail, if you run out of disk space or memory, which
   would lead to a PANIC. Out of disk space could be rather nasty, because
   after WAL replay, we would try to finish the incomplete insertion, which
   would be quite likely to run out of disk space again.

2. The "incomplete actions" mechanism, including the bm_safe_restartpoint()
   rmgr API function, went away in PostgreSQL 9.4. Now that we've merged
   with 9.4, we need to deal with them differently.

After this patch, the insertion is performed in one atomic operation, with
a single WAL record. The XLOG_BITMAP_INSERT_WORDS WAL record format is
changed, so that it can represent the insertion on several bitmap pages in
one record.

Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>
Reviewed-by: BaiShaoqi <sbai@pivotal.io>