Fix the new bitmap AM insertion WAL record.
I botched the WAL-logging of the content words in the
_bitmap_write_bitmapwords_on_page function. In The start position for
changed content words is kept in 'cwords' variable, not 'startWordNo'.
'startWordNo' is the starting position in the input buffer to copy from,
not the position in the target page to copy to.

Move the lines that record this information in the WAL record closer to
the corresponding memcpy()s that make the changes in the master. This makes
it easier to verify that we're recording the same changes in the WAL record
that we are making to the page.

I found this by running a patched version that wrote a full page image
after writing or replaying each XLOG_BITMAP_INSERT_WORDS record, with
the 'bitmap_index' regression test.  This hopefully explains the assertion
failure that Ashwin reported at
https://groups.google.com/a/greenplum.org/forum/#!topic/gpdb-dev/TkoXWveDS6g
although I was not able to reproduce that.