Add WAL debugging code, to dump page contents to a file for comparison.
This is very similar to the "WAL consistency checking" facility added in
PostgreSQL v10. But we don't have that yet, so this will do as a stopgap
until we catch up.

With this patch, a full image of each page by the targeted WAL records
are dumped to a file in the data directory, "bmdump_<segno>_insert". And
at replay, the image after replaying each WAL record is dumped to
"bmdump_<segno>_redo". You can compare the files with each other, to verify
that the WAL redo faithfully reproduces the same page contents as in normal
operation.

This is debugging code that is normally disabled, and needs to be enabled
by the developer if needed.