Add hash opclasses for bit and varbit.
These are missing from the upstream. While not strictly necessary, it seems
weird that these can be used as DISTRIBUTED BY columns, for hash
distribution, but we cannot do hash joins or hash aggregates on them.

My longer-term goal is to get rid of the cdbhash mechanism, and replace
that with regular hash opclasses. Once we do that, we'll need the hash
opclass, or these datatypes can no longer be used in DISTRIBUTED BY. That's
one reason to add these now.