Refactor reshuffle tests, using gp_execute_on_segment() helper function.
This enhances gp_execute_on_segment(), so that it can be used to execute
queries that return a result, and the result is passed back to the caller.
With that, we can avoid the pl/python hack to connect to segments in
utility mode, in reshuffle_ao test.

I imagine that this would be useful for many other tests too, but I
haven't yet searched around for more use cases.

Reviewed-by: Asim R P <apraveen@pivotal.io>
Reviewed-by: Zhenghua Lyu <zlv@pivotal.io>