ExecSupportsMarkRestore Seqscan should return false
This commit fixes the Github issue #6215.

When executing the following plan
```
 Gather Motion 1:1  (slice1; segments: 1)
   ->  Merge Full Join
         ->  Seq Scan on int4_tbl a
         ->  Seq Scan on int4_tbl b
```
Greenplum will raise an Assert Fail. The root cause is
we poorly support mark/restore in `heapam`. We decided
not to touch the code there because later version Postgres
has removed them.

We simply force a material node for the above plan.

Co-authored-by: Shujie Zhang shzhang@pivotal.io
Co-authored-by: Zhenghua Lyu zlv@pivotal.io