Temporarily remove --enable-testutils from targets
The --enable-testutils mechanism was broken in the 9.3 merge, with
fixing it left as a FIXME. Remove the invocation of the option from
gpAux targets as they otherwise fail to compile the tree. This also
expands the comment of the FIXME to indicate that these should be
turned on again when fixed.

Reviewed-by: Jacob Champion <pchampion@pivotal.io>