Don't store total relpages/reltuples in root partition's pg_class.
For consistency with upstream. Sum up the child reltuples at planning
time in ORCA, instead. This adds some overhead to planning partitioned
tables with lots of partitions with ORCA, but we hardly care about the
startup time of ORCA, anyway..

This effectively reverts recent commit 3f3869d627. The tests that were
added to vacuum_gp in that commit don't make much sense anymore, but
instead of removing them altogether, rewrite them into something that is
marginally useful. We probably have tests for reltuples/relpages elsewhere,
in 'analyze' test file at least, but it seems good to have a smaller
summary of the intended behavior, in the form of a test like this.

Reviewed-by: Venkatesh Raghavan <vraghavan@pivotal.io>