Disable deprecated register warning for gpcloud
Including funcapi.h pulls in lwlock.h which use the register storage
class for spinlocks. This causes a compiler warning fo deprecated
storage class when used in C++ code, which gpcloud is.

../../src/include/storage/s_lock.h:220:2: warning: 'register' storage
class specifier is deprecated and incompatible with C++17
[-Wdeprecated-register]
        register slock_t _res = 1;
        ^~~~~~~~~
1 warning generated.

While this will go away as the register stored spinlock is removed
in a future commit which we will get via merging upstream, having a
pointless compiler warning in every buildlog which we all need to
ignore is bad practice so remove it now via a pragma for clang which
is where it was witnessed.