Remove use of xml type in regression test
The xml type is not a default built-in type, but requires the server
to be compiled with --with-libxml. Since the type doesn't seem to do
much in this test, remove it to avoid failing tests in environments
not compiled with libxml support.

Adding an alternate output for builds without XML support would have
solved the issue too, but it would render most, if not all, of this
test useless.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>
Reviewed-by: Jesse Zhang <jzhang@pivotal.io>