Fix test case for changed data distribution with Jump Consistent hash.
Commit 4a174240e6 changed the hash method, which changed the data
distribution of data. It adjusted many test cases that were dependent on
the data distribution, but missed this one. This one isn't failing at the
moment, but it was clearly mean to look at the relfrozenxid of the segment
that it inserts all the data to. Change the value of the distribution key
so that all the data falls into segment 0 again.