In a database-wide ANALYZE, analyze root partitions last.
The HLL method of updating the root stats only works if the leaf partitions
have already been analyzed. So by always analyzing the leafs first, we
can make use of HLL in updating the root stats. Otherwise, it's pure luck
if it works out or not, depending on the order the leafs and parents are
analyzed.