Simplify the data structure to track "too wide" datums slightly.
The per-column 'toowide_cnt' was only used to check if there are any, so
it's redundant with the actual flags. Replace the bool array with
Bitmapset. The Bitmapset is left as NULL when there are no bits set, so we
can use a NULL check in place of checking toowide_cnt.