Use block-level sampling for ANALYZE on heap tables.
The user-visible feature here is to speed up ANALYZE on heap tables, by
using the block-level sampling method from PostgreSQL, instead of the
scanning the whole table. This refactors the ANALYZE code, to make it
possible to call the upstream function to do the block-level sampling
again.

A secondary feature is that this enables the ANALYZE callback for Foreign
Data Wrappers, introduced in PostgreSQL 9.2. It was disabled in the 9.2
merge, because the function signature had been changed in GPDB, making
it cumbersome to call it. This patch reverts acquire-function's signature
to the same as in upstream, so the FDW callback now works again.

This also moves the logic to decide which rels to analyze, based on
optimizer_analyze_root_partition or optimizer_analyze_midlevel_partition
GUCs and the ROOTPARTITION option, wholly to get_rel_oids() in vacuum.c.
It was mostly done there already, but there were some warts in analyze.c.
None of the code in analyze.c should now treat partitions differently
from normal inherited tables, which helps to minimize the diff vs
upstream. (Except that the HLL code still heavily uses partition-related
functions. That's a TODO.)