Explore join order alternatives with LOJ childs
While creating component list to identify the best join order, now childs of
LOJ are considered for possible join combination with other non-LOJ components.
As a result, a transormation like below is possible now (through `Greedy` and
`Mincard`).
```
    Nary           LOJ
     |_t1           |_Nary
     |_LOJ  ==>     | |_t1
       |_t2         | |_t2 (i.e LOJ left child)
       |_t3         |_t3
```
Note: Dynamic Programming xform does not consider explaring alternatives using
LOJ childs.  Also, currently we limit exploring alternatives for join
combination, only the left child of LOJ can be joined with other component of
non-LOJ components.

Consider the below tree:
```
+--CLogicalNAryJoin
	|--CLogicalGet "t1"
	|--CLogicalGet "t2"
	|--CLogicalLeftOuterJoin => LOJ 1
	|  |--CLogicalGet "t3" => {1, EpLeft}
	|  |--CLogicalGet "t4" => {1, EpRight}
	|  +--<Join Condition>
	|--CLogicalLeftOuterJoin => LOJ 2
	|  |--CLogicalGet "t5" => {2, EpLeft}
	|  |--CLogicalGet "t6" => {2, EpRight}
	|  +--<Join Condition>
	+--<Join Condition>
```
While evaluating join order, earlier we were not exploring creating join
combinations of "t1" or "t2" with the outer child of LOJ 1 or 2, i.e "t3" or
"t5". This commit enables considering such alternatives.

The following rules apply while considering the alternatives:
`Rule 1:` joining right child of LOJ with another right child of LOJ
is not a valid combination.
    `example: CLogicalGet "t4" cannot be joined with CLogicalGet "t6"`

`Rule 2:` joining a child of LOJ x with a child of another LOJ y
is not a valid combination, where x and y represent 2 different
LOJs
    `example: CLogicalGet "t3" or "t4" cannot be joined with CLogicalGet "t5" or "t6"`

`Rule 3:` joining right child of LOJ with other NAry component (non LOJs) is not
a valid combination.
    `example: CLogicalGet "t4" or "t6" cannot be joined with CLogicalGet "t1" or "t2"`

`Rule 4:` What is allowed?
a) joining left child of LOJ x with other NAry component (non LOJs) is valid
    `example: joining CLogicalGet "t3" with CLogicalGet "t1" or "t2"`
b) joining left child of LOJ x with the right child of the same LOJ x is valid.
    `example: joining CLogicalGet "t3" with "t4"`
c) joining non LOJ childs is valid, as they are inner joins.
    `example: joining CLogicalGet "t1" with CLogicalGet "t2"`

Note: extending the example to component's containing left child of LOJs The
below component contains left child of LOJ 1
```
+--CLogicalInnerJoin
  |--CLogicalGet "t1"
  |--CLogicalGet "t3" ==> Left child of LOJ 1
  +--<Join Condition>
It can be joined with CLogicalGet "t4", as it ensures that
"t4" remains the right child of the LOJ containing "t3"
```

How it is achived:
- If the flag to include LOJ childs is set, while creating the component list,
  the child of LOJ are also added to the components array
- While creating the component for LOJ child, we track the LOJ to which the
  component belongs, an incremental number is assigned to each LOJ node found
  in the incoming tree
- We also track the position of the LOJ child, i.e whether a component is a
  left or right child of LOJ
- While exploring different combinations, the rules listed above are applied to
  identify if the resulting join in valid.
- If the combination is valid and is cheapest, it is considered.