Fix calculation of AO table's reltuples in new ANALYZE.
The total number of rows in table, stored in pg_class.reltuples, was
calculated incorrectly. It was set to the number of rows in the sample,
not the (estimated?) total number of rows in the table. We didn't notice,
because most of the existing tests use small tables, so that the whole
table fit in the sample. Put back code like we used to have, for
collecting the totals.

This papers over the assertion failure we're seeing in the 'aocs' test in
the concourse pipeline, with ORCA, by changing the plan to one that doesn't
crash. It needs a proper fix, of course, but we're tracking that separately
as https://github.com/greenplum-db/gpdb/issues/6274.