New pg_regress argument to hook the tests.
This is a preparation of an upcoming random ICG pipeline job.  In that
pipeline job we will have tables created on different segments, so far
the only way to hack this behavior is by using the gp_debug_numsegments
extension.  We want to load and execute that extension for all the tests
without modifying them directly, so a hook mechanism is needed.

Added a pg_regress argument --prehook to set a hook.  A hook script
should be put under src/test/regress/{sql,input}/hooks/ directory,
depending on whether it need pg_regress to substitute the @@ tokens.  At
most one hook can be specied.  It can be set via make command like this:

    # suppose there is src/test/regress/sql/hooks/hookname.sql
    make installcheck EXTRA_REGRESS_OPTS=--prehook=hookname