Always create fully distributed tables
In CREATE TABLE we used to decide numsegments from LIKE, INHERITS and
DISTRIBUTED BY clauses.  However we do not want partially distributed
tables to be created by end users, so change the logic to always create
tables with DEFAULT as numsegments.  We still allow developers to hack
the DEFAULT numsegments with the gp_debug_numsegments extension.