Stop abusing Result's hash filter for running a plan on arbitrary segment.
ORCA generated plans, where the "hash filter" in the Result node was set
to an empty set of columns. That meant "discard all the rows, on all
segments, except one segment". This is used at least with set-returning
functions, where we don't care where the function is executed, but it only
needs to be executed once. (The planner creates a one-to-many Redistribute
Motion plan in that scenario, which makes a lot more sense to me, but
doing the same in ORCA would require more invasive surgery than what I'm
capable of.)

Instead of executing the subplan, and throwing away the result one row at
a time, use a Result plan with a One-Off Filter. That's more efficient.
Also, it allows removing the Result.hashFilter boolean flag, because there
the weird case of a hashFilter with zero columns is gone. You can check
"hashList != NIL" directly now.

The old method would always choose the same segment, which seems bad for
load distribution. The way it was chosen seemed totally accidental too:
we initialized the cdbhash object to the initial constant value, and
then reduced that into the target segment number, using the jump
consistent hash algorithm. We computed that for every row, but the result
was always the same. On a three-node cluster, the target was always
segment 1. Now, we pick a segment at random when generating the plan.

Reviewed-by: Melanie Plageman <mplageman@pivotal.io>