Include binary coercible casts for predicate inference
This commit augments ORCA's predicate inference system, by also
considering binary coercible casted scalar idents & constants, both on
the lhs and rhs of an expression. The logic strips the cast and uses
the underlying CColRef in equiv classes. It also passes the CColRef to
the constraint framework. This way the logic in in the constraint
framework remains unchanged.

Note, however, that when the inferred constraint is converted back to an
expression, appropriate casts (binary coercible or otherwise) must be
added back. This is handled in CUtils::PexprScalarCmp(). Once we
retrieve the appropriate operator to use for construct the scalar cmp,
we add necessary casts to both the left and right expressions,
according to the left & right oids required by the scalar cmp operator.

Also, refactor the code in the adjacent areas:

- This commits removes a lot of the duplicate code in methods:
  CUtils::GetScCmpMdId(), CUtils::FCmpOrCastedCmpExists, BOOL
  FCmpExists() and CUtils::PexprScalarCmp(). The order of checks in
  these methods was different, for no apparent reason. Thus, this
  refactor helps bring similar functionality that is easier to maintain.

- Move similar functions to CMDAccessorUtils

- Apply correct casts to match operand mdids for ScCmp & IDF
  GetScCmpMdIdConsiderCasts() returns the mdid of the operator to use to
  construct the scalar cmp op, but doesn't modify the expressions. This
  commit adds a new method GetScCmpMdIdApplyCasts() that determines the
  correct cast(s), if any, to add on the left and right expressions. It
  does this by comparing the mdids of the expressions and the input
  mdids expected by the operator.

  For example, when asked for =(varchar, varchar), if the op returned by
  GetScCmpMdIdConsiderCasts() is =(text, text), both the lhs and rhs
  must be cast to text.

Finally, add new MDPs for inferred predicates

For now with constant expression evaluation off, no predicate inferring
is happening. These can be update once it is enabled.

Bump ORCA version to 3.16

Co-authored-by: Hans Zeller <hzeller@pivotal.io>