Update centos6 build jobs in gpdb_master to use new image
In order to provide better tracking and management of build-time
dependencies for GPDB6, a new Docker image is being produced in the
gp-image-baking pipeline. This commit updates the enterprise build jobs
in the master pipeline to reference the new container image; the OSS
build jobs are not modified. This change is in preparation for the
creation of the 6X_STABLE branch.

The Docker images used to build enterprise GPDB for other platforms will
be updated in subsequent commits.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>
Co-authored-by: Taylor Vesely <tvesely@pivotal.io>