New pipeline job to run ICG with random numsegments.
This new pipeline job run ICG tests, all the tables are created with
random numsegments, the purpose is to improve test coverage on MxN
queries.

Some tests are disabled because their behavior and/or output depend on
data distribution, in general could be grouped as below:
- checks gp_segment_id;
- contains EXCHANGE PARTITION, which require the two tables have the
  same numsegments;
- turns on `test_print_direct_dispatch_info`, the output depends on
  numsegments;

Some tests can pass by adding some enforcement or alignments on
numsegments, but we do not want to include such kind of hacks on those
tests for now.

Some tests can be re-enabled later if we can make them stable without
changing them.