Suppress a warning generated on inherited tables.
The following WARNING is generated by ANALYZE when some sample tuples
are from segments outside the [0, numsegments-1] range, however this
does not indicate the data distribution is wrong.  Take inherited tables
for example, when inherited tables has greater numsegments than parent
this WARNING will be raised, and it is expected.  This could happen
normally on the random_numsegments pipeline job, so ignore this WARNING.

    WARNING:  table "patest0" contains rows in segment 2,
              which is outside the # of segments for the table's policy
              (2 segments)

Added this pattern to init_file to ignore it.