Fix OOM during HLL merge calculation
We need to free up the temporary HLL counters created during HLL merging,
if there are lot of partitions and the data is high for each partition,
we may run into OOM if the temporary counters are not released.

In the merge_leaf_stats function, a nested for loop leaks a single HLL
counter for N^2 iterations (where N = number of partitions). This is run
in a single context, hence leaking considerable memory, leading to VMEM
errors.

This patchs fixes the leaks.

Co-authored-by: Rahul Iyer <riyer@pivotal.io>
Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>