Fix numeric round() and trunc() to not scribble on their input.
numeric_round() and numeric_trunc() initialized the NumericVar to
represent the argument with init_var_from_num(), and then called
round_var() on it. That's not cool, init_var_from_num() creates a
read-only NumericVar, whose 'digits' points directly to the argument
datum, and round_var() scribbles on it. There is a comment in
init_var_from_num() warning explicitly to not do exactly that, but we
failed to heed the warning.

Fixes github issue https://github.com/greenplum-db/gpdb/issues/6383. I
also added a test case for this, although it's a bit hard to see this bug
to reappear in exactly the same form. It's not far-fetched that we might
re-introduce it in a slightly different form, though, so I also added some
extra Asserts in round_var() and trunc_var() for it.

Reviewed-by: Ekta Khanna <ekhanna@pivotal.io>