Avoid including 1 kB of zeros in every QD -> QE query dispatch.
The 'M' type QD->QE dispatch message includes a serialized version of the
current resource group information. That included a 'cpuset' field, which
has 1 kB of space reserved for it, and the serialization routine included
all the padding zeros in it. (Copying or memsetting that into a local
variable in SerializeResGroupInfo() wasn't completely free either).
Rewrite the serialized format to be less silly.

Reviewed-by: Jacob Champion <pchampion@pivotal.io>