Remove SwitchedMemoryContext
The SwitchedMemoryContext is not an actual new context type, but
merely a set of convenience functions for wrapping allocation and
switching into a single call. This was only used at two callsites,
and cause pointless merge conflicts with upstream (not to mention
making code reading harder since it's a new concept from what we're
all used to). This removes and replaces with normal memory context
calls.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>
Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>