Remove convenience function for TopMemoryContext dumping
The dump_tmc() function is just a shorthand for running dump_mc()
on the TopMemoryContext. Since dumping MemoryContexts has no call-
sites in the code and is only available in a debugger, it seems a
bit rich to spend object file space for saving a few keystrokes in
GDB. This removes dump_tmc() in favour of just using dump_mc().

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>
Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>