Remove unused 'consec_csv_err' code.
The 9.1 merge removed the code that incremented 'num_consec_csv_err'.
Since then, it was only ever initialized to 0. Hence, all the related code
was dead.

I'm not sure how we behave with the kinds of errors that used to trigger
this code now. But I don't see a need to treat them specially, the
generic error handling code should cope with them. The GUC that controlled
the max CSV line length was removed in commit 6ac29fdc32 already.