Fix crash with 'gp_enable_motion_mk_sort=off'.
The non-MK code was abusing the child's result tuple slot. The
corresponding MK code was changed back in 2010 to not do that, but that
commit missed the non-MK version. This caused a segfault in the
'subselect' regression test. Apparently no one has run the regression
tests with 'gp_enable_motion_mk_sort=off' recently.

Reviewed-by: Pengzhou Tang <ptang@pivotal.io>
Reviewed-by: Gang Xiong <gxiong@pivotal.io>