gpexpand improvement using newly added features
commit 8eed421738 & e0b06678aa allow us to do a gpexpand of GPDB cluster
without a restart, so we called this strategy as "online expand", those
two commits were mainly focus on how to avoid restarting cluster when
expanding cluster, this commit will do the left work to improve gpexpand
using a few features we merged to master recently.

First improvement is, it's no longer necessary to change policy of all
non-partition table to random at the beginning of gpexpand. Previously,
we couldn't tell the difference between expanded table and non-expaned
table, so if the distribution policy of these tables were the same,
planner took these tables's data as co-located and produced a incorrect
plan. To avoid this, gpexpand used to change policy of all table to
random so planner could produce a correct but a non-effective plan,
because random policy always means annoying broadcase motion.

Now each table has a 'numsegments' attribute introduced by 4eb65a53, GPDB
can recognize expanded and non-expand table and produced correct plans. so
the first imporvement is removing the randomization step of tables.

The second improvement is, use a brand new syntax "ALTER TABLE foo EXPAND TABLE"
to focus on the data rebalance of tables. Previously, tables were converted
to randomly before rebalance and numsegments of tables were always the same
with GPDB cluster size, so we can use a tricky "ALTER TABLE foo SET WITH
(REORGNIZE = true) DISTRIBUTED BY (original_key)" command to rebalance data
to new added segments, now policy and numsegments are not changed before
rebalancing data, so expanding such tables is not match the concept of "SET
DISTRIBUTED BY" now, we need a proper syntax to focus on table expanding.

New expand syntax named "ALTER TABLE foo EXPAND TABLE", we have two methods
to do the real data movement inside, one is CTAS, another is RESHUFFLE, which
one is better depends on how much data needs to be moved and whether the table
has index and whether the table is a append only table (analyzed by Hekki).

A drawback of this commit is we always expand a partition within a transaction,
the old behavior is expand the leaf partition in parallel which is faster for
a partition table. We don't allow root partition and its leaf partitions to
have different numsegments for now, this commit disabled the old behavior
temporarily. A topic on how to expanding partition table in parallel is
under discussion and we wish to bring the ability back properly in the feature.

* Refine name quote

Eg: if schema name of table is a.b, table name is c'd, then current gpexpand
cann't handle it.