Remove leftover code for non-existent 'G' and 'W' protocol messages.
Looking at the historical pre-open-sourcing repository, the 'G' message
was used by ancient prototype code from 2007-2009. The prototype code was
removed in 2009, but this snippet was left over. I don't think it was used
in any released version of GPDB.

The 'W' type message was removed in commit daf6cdbc67, but this was left
over.