Optimize CHECK_FOR_INTERRUPTS()
CHECK_FOR_INTERRUPTS() had grown quite bloated, given how frequently it runs.
It called IsResQueueEnabled(), which added the overhead of a function call,
even if resource queues were not enabled. If they were enabled, it would
call BackoffBackendTick(), which adds another function call.
BackoffBackendTick() checked a lot of other conditions, so it would still
often not do anything.

It turns out to be cheaper to always increment the "tick" counter, to avoid
the overhead of checking whether we need it or not. Move all the the checks
for whether the whole backoff mechanism is even needed, into a function
that only runs when the tick counter expires.

I've seen those calls take up to 5% of CPU overhead, when profiling some
queries with linux 'perf'. This should shave off the time spent on that.

Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>
Reviewed-by: Ning Yu <nyu@pivotal.io>