Optimize the code to send data rows from QD to QEs in COPY FROM.
This code is performance-critical, as the QD will quickly become the
bottleneck when loading data with COPY FROM. Optimizing it is therefore
warranted.

I did some quick testing on my laptop, loading a table with 50 boolean
columns. This commit improved the performance of that by about 10%.

In GPDB 5, we used to postpone calling the input functions for columns
that were not needed to determine which segment to send the row to. We
lost that optimization during the 9.1 merge, when the COPY code was
heavily refactored. We now call the input functions for every column in
the QD, which added a lot of overhead to the QD. On the other hand, some
other things got faster. Before this commit, the test case with 50
booleans was about 10% slower in GPDB 6, compared to GPDB 5, so this buys
back the performance, so that it's about the same speed again. I'm sure
you can find a test case where GPDB 6 is still slower - you can make the
input functions arbitrarily slow, after all - but every little helps.

Reviewed-by: Adam Lee <ali@pivotal.io>