Foreign data wrapper and server support mpp_execute option
The values would be the default values for things of them.

For example:

If a foreign table has not set the mpp_execute, it follows the server,
if a server has not set the mpp_execute, it follows the wrapper,
if a wrapper has not set the mpp_execute, it chooses the default value,
which is 'master'.