Avoid skew in motions for intermediate aggregate
When three stage aggs are produced, the optimizer considers two
alternatives while adding motions. One is with a motion on top of the
intermediate agg which reduces the skew, and the other is without a
motion but possible producing skew based on the data. This patch takes a
conservative approach and disables the second alternative so that the
plan produced will always have an extra motion and reduce the possible
skew.

This patch also introduces a guc `optimizer_enable_agg_skew_avoidance`
which can be used to turn on or off the behavior. This guc is set to on
by default which means that the second alternative will be disabled.

Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>
Co-authored-by: Hans Zeller <hzeller@pivotal.io>