resgroup: go back to old behavior when spill memory is zero.
When spill memory is zero the original behavior is to use work memory
instead, however we modified this for resource group to use dynamic
spill memory in 98f9b1744ffc8d6383b7bdcaee80fb2f2d26cae1.  This is not
always the expected behavior, especially when a warning is generated.

Change back to old behavior.