Improve performance of `merge_leaf_stats()`
The previous algorithm to merge the HLL counters for leaf partitions had a time
complexity of O(n^2) where n is the number of leaf partitions. In case where
the number of partitions was high, resulting in poor performance of analyze on
partitioned tables.  This patch fixes the performance issue by using a new
algorithm to merge the leaf HLL counters in O(n) time.

Co-authored-by: Omer Arap <oarap@pivotal.io>
Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>