Refactor executor code for TableScan, DynamicTableScan, BitmapHeapScan.
This removes a lot of GPDB-specific code that was used to deal with
dynamic scans, and code duplication between nodes dealing with Heap, AO
and AOCS tables.

* Resurrect SeqScan node. We had replaced it with TableScan in GPDB.
  Teach SeqScan to also work on append-only and AOCS tables, and remove
  TableScan and all the code changes that were made in GPDB earlier to
  deal with all three table types.

* Merge BitmapHeapScan, BitmapAppendOnlyScan, and BitmapTableScan node
  types. They're all BitmapHeapScans now. We used to use BitmapTableScans
  in ORCA-generated plans, and BitmapHeapScan/BitmapAppendOnlyScan in
  planner-generated plans, and there was no good reason for the
  difference. The "heap" part in the name is a bit misleading, but I
  prefer to keep the upstream name, even though it now handles AO tables
  as well. It's more like the old BitmapTableScan now, which also handled
  all three table types, but the code is refactored to stay as close to
  upstream as possible.

* Introduce DynamicBitmapHeapScan. BitmapTableScan used to perform Dynamic
  scans too, now it's the responsibility of the new DynamicBitmapHeapScan
  plan node, just like we have DynamicTableScan and DynamicIndexScan as
  wrappers around SeqScand and IndexScans.

* Get rid of BitmapAppendOnlyPath in the planner, too. Use BitmapHeapPath
  also for AO tables.

* Refactor the way Dynamic Table Scan works. A Dynamic Table Scan node
  is now just a thin wrapper around SeqScan. It initializes a new
  SeqScan executor node for every partition, and lets it do the actual
  scanning. It now works the same way that I refactored Dynamic Index
  Scans to work in commit 198f701ede. This allowed removing a lot of code
  that we used to use for both Dynamic Index Scans and Dynamic Table
  Scans, but is no longer used.

There's now some duplication in the Dynamic* nodes, to walk through the
partitions. They all have a function called setPidIndex(), for example,
which does the same thing. But I think it's much more clear this way,
than the previous DynamicController stuff. We could perhaps extract some
of the code to common helper functions, but I think this is OK for now.

This also fixes issue #6274. I'm not sure what exactly the bug was, but it
was clearly in the Bitmap Table Scan code that is used with ORCA-generated
plans. Now that we use the same code for plans generated with the Postgres
planner and ORCA, it's not surprising that the bug is gone.

Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>
Reviewed-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>